requirements.txt
yfinance
pandas
numpy
requests
urllib3
python-dotenv
feedparser
holidays
anthropic
